requirements.txt
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.2.0
cachelib==0.1.1
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
click==7.1.2
Cython==0.29.14
DateTime==4.3
decorator==4.4.2
entrypoints==0.3
Flask==1.1.2
Flask-Session==0.3.2
fuzzywuzzy==0.18.0
google-api-core==1.22.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.30.0
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
importlib-metadata==1.7.0
iniconfig==1.0.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
msgpack==1.0.0
neovim==0.3.1
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.3.0.38
packaging==20.4
pandas==1.0.3
parso==0.7.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==3.0.4
protobuf==3.12.2
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pynvim==0.4.1
pyparsing==2.4.7
pystan==2.19.1.1
pytesseract==0.3.4
pytest==6.0.0
pytest-flask==1.0.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
pyzmq==18.1.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.0
six==1.14.0
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
twilio==6.40.0
typed-ast==1.4.1
urllib3==1.25.9
waitress==1.4.4
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
